+++ crates/core/src/pattern_compiler/mod.rs
@@ -32,6 +32,7 @@ pub(crate) mod pattern_definition_compiler;
 pub(crate) mod predicate_compiler;
 pub(crate) mod predicate_definition_compiler;
 pub(crate) mod predicate_return_compiler;
+pub(crate) mod rewrite_compiler;
 pub(crate) mod step_compiler;
 
 pub(crate) use compiler::{parse_one, CompilationContext};
